+++ rust/pe010/src/main.rs
@@ -1,6 +1,6 @@
-fn is_prime(test: u64, primes: &Vec<u64>) -> bool {
+fn is_prime(test: usize, primes: &Vec<usize>) -> bool {
     let root = (test as f64).sqrt();
-    let introot = root.floor() as u64;
+    let introot = root.floor() as usize;
 
     for i in primes {
         if i > &introot {
@@ -13,15 +13,15 @@ fn is_prime(test: u64, primes: &Vec<u64>) -> bool {
     return true
 }
 
-fn pe010() -> u64 {
-    let last:u64 = 2000000;
+fn pe010() -> usize {
+    let last:usize = 2000000;
     let mut primes = vec![2];
 
-    let mut psum:u64 = 2;
-    let mut test:u64 = 3;
+    let mut psum:usize = 2;
+    let mut test:usize = 3;
     while test < last {
         if is_prime(test, &primes) {
-            primes.push(test as u64);
+            primes.push(test as usize);
             psum += test;
         }
         test += 2;
